mt5_collector/requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0
pydantic>=2.0
httpx>=0.24
APScheduler>=3.10
python-dotenv>=1.0
